+++ database/migrations/2025_01_17_094545_create_coupons_table.php
@@ -13,21 +13,23 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('coupons', function (Blueprint $table) {
-            $table->id();
-            $table->string('code')->unique();
-            $table->text('description')->nullable();
-            $table->enum('type', ['percentage', 'fixed']);
-            $table->decimal('value', 10, 2);
-            $table->decimal('min_order_value', 10, 2)->nullable();
-            $table->datetime('start_date');
-            $table->datetime('end_date');
-            $table->integer('max_uses')->default(0);
-            $table->integer('uses_per_user')->default(0);
-            $table->boolean('status')->default(1);
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (! Schema::hasTable('coupons')) {
+            Schema::create('coupons', function (Blueprint $table) {
+                $table->id();
+                $table->string('code')->unique();
+                $table->text('description')->nullable();
+                $table->enum('type', ['percentage', 'fixed']);
+                $table->decimal('value', 10, 2);
+                $table->decimal('min_order_value', 10, 2)->nullable();
+                $table->dateTime('start_date');
+                $table->dateTime('end_date');
+                $table->integer('max_uses')->default(0);
+                $table->integer('uses_per_user')->default(0);
+                $table->boolean('status')->default(1);
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
@@ -37,6 +39,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('coupons');
+        if (Schema::hasTable('coupons')) {
+            Schema::dropIfExists('coupons');
+        }
     }
 };
